Tensorflow CNN MNIST training
To create an end-to-end convolutional neural network (CNN) in
TensorFlow that will run on the host CPU system. It will have several
macro parameters that can be manually adjusted in the files to change
the accuracy and the final image to be analyzed. The macro parameters
will be indicated in the activities. The files are a mixture of bash and
python3 files, with the graphics at the end using the matplotlib module.
This activity should produce a baseline for comparison to other systems
e.g. CPU, Movidius MYRIAD 2, and OpenVino supported systems.